Simplify framework with utils and remove useless registry
